moved tests into a separate function, tweaked quicksort sorting
