+++ MostCommonElementInArray/MostCommonElementInArray/MostCommonElementInArray.h
@@ -1,8 +1,4 @@
 #pragma once
-#include <stdbool.h>
 
 // function to find the most frequent element in an array
 int searchMostCommonElementInArray(int* arrayOfNumber, int numberOfElements);
-
-// test for function searchMostCommonElementInArray
-bool testSearchMostCommonElement();
