+++ MostCommonElementInArray/MostCommonElementInArray/MostCommonElementInArray.vcxproj
@@ -142,10 +142,12 @@
     <ClCompile Include="main.c" />
     <ClCompile Include="MostCommonElementInArray.c" />
     <ClCompile Include="qsort.c" />
+    <ClCompile Include="test.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="MostCommonElementInArray.h" />
     <ClInclude Include="qsort.h" />
+    <ClInclude Include="test.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
