+++ MostCommonElementInArray/MostCommonElementInArray/MostCommonElementInArray.vcxproj.filters
@@ -24,12 +24,18 @@
     <ClCompile Include="main.c">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
+    <ClCompile Include="test.c">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="qsort.h">
       <Filter>Файлы заголовков</Filter>
     </ClInclude>
     <ClInclude Include="MostCommonElementInArray.h">
+      <Filter>Файлы заголовков</Filter>
+    </ClInclude>
+    <ClInclude Include="test.h">
       <Filter>Файлы заголовков</Filter>
     </ClInclude>
   </ItemGroup>
