+++ MostCommonElementInArray/MostCommonElementInArray/test.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <stdbool.h>
+
+// function for testing sorting QSort
+bool testCorrectQSort();
+
+// function for checking the sorting of an array
+bool sortingCheck(int* arrayOfNumber, int numberOfElements);
+
+// test for function searchMostCommonElementInArray
+bool testSearchMostCommonElement();
+
